requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
distro==1.9.0
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
gradio==5.49.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-core==0.3.78
langchain-huggingface==0.3.1
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.32
MarkupSafe==3.0.3
narwhals==2.7.0
numpy==2.3.3
openai==2.2.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.32.1
pyarrow==21.0.0
pydantic==2.11.10
pydantic_core==2.33.2
pydeck==0.9.1
pypdf==6.1.1
PyMuPDF==1.26.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
streamlit==1.50.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
zstandard==0.25.0
